+++ crates/nbcl_impl/src/libraries/mod.rs
@@ -1,7 +1,4 @@
-mod apilib;
 mod corelib;
-
-pub use apilib::*;
 pub use corelib::*;
 
 #[macro_export]